# PYTHON/app/schema/auth.py
from pydantic import BaseModel
from typing import Optional
from datetime import datetime

class UserBase(BaseModel):
    username: str

class UserCreate(UserBase):
    password: str

class User(UserBase):
    id: int
    created_at: datetime
    updated_at: Optional[datetime] = None

    class Config:
        orm_mode = True

class Login(BaseModel):
    username: str
    password: str

class TokenRefresh(BaseModel):
    refresh_token: str